Add script to wait for Consul to be available
This allows us to tolerate a 10 minute delay between when we spin up the
service and when we grant firewall access to Consul, which seems
reasonable.

Eventually, this kind of thing would be its own separate "agent" that
comes with its own unit tests, but for now this is just an example of
how to glue things together with Cloudless.